+++ js_libraries/types/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.5.0
 - Add a new `flowIds` subfield to the TraceOption parameter in the Trace API to support one-to-many trace associations.
+- Refactor event definitions for improved IDE compatibility.
 
 ## 3.4.9
 - Add more interface for `scroll-view`.